Release to pre prod

+++ NMP-Portal/NMP-Commons/Models/OrganicManure.cs
@@ -19,7 +19,7 @@ public class OrganicManure
     public DateTime EndOfDrain { get; set; }
     public int Rainfall { get; set; }
     public decimal? AreaSpread { get; set; }
-    public decimal? ManureQuantity { get; set; }
+    public int? ManureQuantity { get; set; }
     public int? ApplicationMethodID { get; set; }
     public int? IncorporationMethodID { get; set; }
     public int? IncorporationDelayID { get; set; }
